terminal.integrated.cwd: Remove pointless setting.
No idea why I set this to ${workspaceFolderBasename}, of course this is
bullshit.
What I wanted was ${fileDirname} to always open the terminal in the
currently opened file's dir.
This is useful for the "scripts" workspace, but not for projects that
need to be compiled.
Furthermore, I can open a terminal in the currently opened file's dir
via ctrl+T.
Thus, I reset this setting in the User config and will set it to
${fileDirname} in the workspace settings of "scripts".