Set "window.zoomLevel" to -0.1 (for TravelMate2 only).
This makes everything (e. g. fonts and side bar icons) a little bit smaller,
so an additional icon fits in the side bar and the number of columns and
lines in editor and terminal windows is increased.

Thanks to the high pixel density of TravelMate2's display, I can still
read everything without stressing my eyes.
However, this value is the lowest limit for me.